Fine tuned Standalone Select Queries to use proper column names.
- Moved direct select tests from `GeneralTranslationTests.cs` to new `DirectSelectCallTests.cs`.
- Updated `LambdaExpressionConverter` to ignore child ParameterExpression conversions.
- Updated `MemberExpressionConverter` to wrap the query if created directly through Select.
- Updated `ParameterExpressionConverter` to wrap the query if created directly through Select.
- Improved projection handling in `StandaloneSelectQueryMethodExpressionConverter`.
- Added validation in `SqlQueryExpression` to ensure correct projection types when creating `SqlQueryExpression` using direct Select.
