+++ src/Atis.SqlExpressionEngine.UnitTest/Tests/FullDataSourceSelectionTests.cs
@@ -17,7 +17,7 @@ namespace Atis.SqlExpressionEngine.UnitTest.Tests
             Expression<Func<object>> queryExpression = () =>
             dbc.DataSet<Student>()
             .Where(x => x.StudentId == "123")
-            .Select(x => x);
+            .Select(f => f);
             string? expectedResult = @"
 select	a_1.StudentId as StudentId, a_1.Name as Name, a_1.Address as Address, a_1.Age as Age, a_1.AdmissionDate as AdmissionDate, a_1.RecordCreateDate as RecordCreateDate, a_1.RecordUpdateDate as RecordUpdateDate, a_1.StudentType as StudentType, a_1.CountryID as CountryID, a_1.HasScholarship as HasScholarship
 	from	Student as a_1
