+++ src/Atis.SqlExpressionEngine/ExpressionConverters/ParameterExpressionConverter.cs
@@ -83,7 +83,7 @@ namespace Atis.SqlExpressionEngine.ExpressionConverters
         {
             var sqlExpression = this.GetDataSourceByParameterExpression()
                             ??
-                            throw new InvalidOperationException($"No SqlExpression found for ParameterExpression '{this.Expression}'. This error usually indicates that the Query Method converter is not converting the first parameter to SqlQueryExpression, e.g. Select(customExpression, x => x.Field1), assume that 'customExpression' is not converting correctly to SqlQueryExpression, therefore, parameter 'x' will not be linked to any SqlQueryExpression instance and will cause this error when translating 'x' part of 'x.Field1' expression. Another reason could be that a custom query method's LambdaExpression parameter is not being mapped to any Data Source. E.g. CustomMethod(query, x => x.Field1, (p1, p2) => new {{ p1, p2 }}), so 'p1' and 'p2' parameters might be presenting data sources but not mapped to any. This is the responsibility of CustomMethod converter class to map those using ILambdaParameterToDataSourceMapper.");
+                            throw new InvalidOperationException($"No SqlExpression found for ParameterExpression '{this.Expression}'. This error usually indicates that the Query Method converter is not converting the first parameter to SqlQueryExpression, e.g. Select(customExpression, x => x.Field1), assume that 'customExpression' is not converting correctly to SqlQueryExpression, therefore, parameter 'x' will not be linked to any SqlQueryExpression instance and will cause this error when translating 'x' part of 'x.Field1' expression. Another reason could be that a custom query method's LambdaExpression parameter is not being mapped to any Data Source. E.g. CustomMethod(query, x => x.Field1, (p1, p2) => new {{ p1, p2 }}), so 'p1' and 'p2' parameters might be presenting data sources but not mapped to any. This is the responsibility of CustomMethod converter class to map those using {nameof(ILambdaParameterToDataSourceMapper)}.");
             var isLeafNode = !(this.ParentExpression is MemberExpression);
             // isLeafNode is true when the ParameterExpression is selected alone
             if (!(sqlExpression is SqlQueryExpression || sqlExpression is SqlDataSourceExpression))
@@ -100,6 +100,13 @@ namespace Atis.SqlExpressionEngine.ExpressionConverters
             {
                 if (sqlExpression is SqlQueryExpression sqlQuery && !sqlQuery.IsMultiDataSourceQuery)
                 {
+                    if (sqlQuery.InitialDataSource == null) 
+                    {
+                        // this would the case where a query is created without From i.e. direct Select call
+                        // in that case we need to wrap the query so that columns will be available to select
+                        sqlQuery.WrapInSubQuery();
+                    }
+
                     var scalarVal = sqlQuery.GetScalarColumnExpression();
                     if (scalarVal != null)
                         return scalarVal;
